Update patch set 1
No score.

Well, I think - it is not very good idea to make such symlink, cause
only root is allowed to change something in /bin. Of course, most of
potential SDK users usually are superusers of host system, but not
everyone. Other objection is - IMHO, bash is not in core utils and it's
place in /usr/local, so better not to change this behaviour.

PS:
however, making symlink solves many problems without additional changes
to sources :)

Patch-set: 1