Update patch set 1
I would prefer that you didn't submit this.

I don't agree with this change.  In some cases you don't want deprecated
to pass through, but in most you do.  We have the @undeprecate tag to
get the behavior that this change does everywhere.

For example, if we deprecate an obscure method on View, we shouldn't
have to go through the entire API and override that method on every
class that indirectly inherits from View and also mark it deprecated.

Patch-set: 1