Update patch set 1
Patch Set 1: No score; No score

I'm not sure that those are equivalent at the low level (but I'm neither a sed not a tr expert). It seems to me that the original sed version replaces each instance of the string "  *" with a newline, whereas the tr version replaces each individual space and star with a newline.

Patch-set: 1
Label: Verified=0