Update patch set 1
Patch Set 1: No score; No score

Are you actually seeing a problem, here?  $@ is defined to exclude the name of the executable; it seems to replace the first element of $@ with an empty string.

Your change is arguably more correct, though, even if it doesn't seem to change the output for me.

Patch-set: 1