Update patch set 1
Patch Set 1: No score; Looks good to me, but someone else must approve

This looks generally good to me, but it would be nice to get Ben Leslie's take on it (since he added the original armv4 support).

Ping me if you don't hear back from him in a few days, and I'll assume it's ok and approve it.

Patch-set: 1