Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

The design of autoconf is fundamentally deeply different from that of the android build system (recursive vs. non-recursive, configuration mechanism, cleanup mechanisms, location of the output), which could cause some confusion with bad consequences.

Worse, the build system is critically at the center of the Android Open-Source Project, and this contribution would be fundamentally unmaintained (since it'd be unused in the Android Open-Source Project itself), with unpredictable consequences.

Because of those risks, autoconf support doesn't make sense in Android, and if a module really wants to use it it should use a local copy of those build rules (and maintain that local copy as the build system evolves).

Patch-set: 1
Reviewer: Gerrit User 1000660 <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0