Update patch set 1
As far as I'm concerned 2% is acceptable, when compared to the trickiness of getting an old Java running on a recent OS, or the annoyance of getting a broken build because of a non-Java-1.5-compatible change, or compared to the various tricks that can be used to make builds faster.

It's only for a finite period of time, to allow people to transition their systems to 1.6. It'll only be used for a few months, and at some point once everyone runs 1.6 we can disable it (as well as the 1.5 enforcement in older branches).

Patch-set: 1