Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The change does merge cleanly on top of master. However, like Joe, I'm questioning the value of this change. The need to version a library sounds very remote (why not have a library that implements all the necessary functionality?), and the need to version so many libraries that we need to support that in the build system is even more remote.

Patch-set: 2
Label: Code-Review=-1