Update patch set 2
Patch Set 2:

The whole point of inheritance is precisely to not need to worry about implementation details about what one inherits from, and this change exists for the sole purpose of depending on those implementation details.

I think that it should be possible to re-write inheritance hierarchies to not require such peeking under the covers.

Patch-set: 2