Update patch set 4
Patch Set 4: No score; No score

I've only looked at the stuff in tools/releasetools; not the rest of the build system changes.

At a minimum the use of the retouch feature needs to be controlled by a flag to ota_from_target_files, and that flag should default to off.  We may enable it in the future, but I'm not at all convinced that we want to use this.

Patch-set: 4
Label: Code-Review=0
Label: Verified=0