Update patch set 4
Patch Set 4:

If gzip works for you, I'd rather just say the gzip is the format and switch everyone to that.  I'm sure that someone will say bzip is better in some way, but given that android uses gzip pretty much everywhere else I think we should go with that, especially because we include zlib all over the place.

Brian and Dima, do you care to comment?

Patch-set: 4