Update patch set 1
Patch Set 1:

Comments inline

  Comments from: Joe Onorato
  TARGET_ARCH should be figured out based on which product is being built,
  or more properly which board is being used. It should not be set by
  lunch or buildspec.mk directly. It doesn't make sense to build the arm
  products with x86 or vice versa.

I would also prefer to set TARGET_ARCH in BoardConfig.mk.
If this works, we should be able to never change the buildspec.mk file and
should also be able to do
 lunch generic-eng
 make
 lunch generic_x86-eng
 make
Is this what you had in mind?


  Comments from: Joe Onorato
  I don't mind having the generic targets hard coded where they are. How
  many more architectures are there beyond arm and x86 that we're likely
  to ever support? I don't think any of them would probably need generic
  targets to be in the lunch menu either. That list should be kept short
  anyway.

Ok... I can move the generic_x86 lunch target to build/envsetup.mk. I really
think that it would be better to move generic OUT of that file though.

Patch-set: 1