Update patch set 1
Patch Set 1: Verified; I would prefer that you didn't submit this

Thanks for the contribution.

core.mk and generic.mk are used as the base for every device build, which means that applications that are added there must be vital to every single device. Given the number of devices that have shipped so far without such wap push support, I'm not convinced that the wap push manager qualifies.

It might make more sense to add it to full.mk instead, which is meant to contain every possible application, though not necessarily in a combination that is used to ship a device.

I'll let the build experts decide.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=+1