Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

First there is a problem in the change:
If a function is redefined in more than 1 vendor definitions.mk, then which definition will be used is not deterministic.

I am not sure why do we need such a mechanism? if you want to change the global behavior of a function, just change its definition in the global definitions.mk; if you want to change the behavior specific to some module, you can add new function or rule in the global makefile.

Patch-set: 1
Reviewer: Gerrit User 1003981 <1003981@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0