Update patch set 1
Patch Set 1: Do not submit

Yeah, I agree with Ying.  The definitions.mk file is the very core of the build system.  I don't like the idea of adding a facility for it to be modified based on which vendor trees are installed.

There is probably a better way of factoring out whatever you're trying to do.

Patch-set: 1
Reviewer: Gerrit User 1000701 <1000701@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0