Update patch set 4
Patch Set 4: Verified; Looks good to me, but someone else must approve

I verified it by performing a clean build of full_x86-eng, which seems to work perfectly fine under emulation.

For the record, what we really want is to have the least patches between upstream and our own toolchain repository. This implies that we want to send patches upstream as soon as possible and work to get them accepted if possible.

The ARM toolchain was in a similar position not so long ago. It is only very recently that enough ARM/Android-related patches have been integrated into upstream.

I expect the IA/Android situation to be similar, i.e. we're starting with lots of explicit compiler flags, then provide the appropriate configuration files in the gcc tree to make an "arm-linux-android-gcc" be a proper driver that doesn't need special flags to be used normally.

I don't have +2 on the build tree though.

Patch-set: 4
Reviewer: Gerrit User 1000411 <1000411@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=+1