Update patch set 1
Patch Set 1: Do not submit

ok, I think I understand. This patch comes from a series of build/ changes that happened in AOSP and which predate some of the x86-support ones.

Looking at the patch's content, I don't think we should back-port this one. This doesn't seem to affect the rest of the x86 patches after this one (I looked them up quickly), and does affect the ARM builds in significant ways.

So I'm putting a -2 on this one, and will ignore it in my testing. If a problem arises when trying to cherry-pick a future backport patch, I'll report it and try to fix the issue.

Patch-set: 1
Label: Code-Review=-2