Update patch set 3
Patch Set 3:

> This doesn't make sense.

I agree.  But I am confident this is what is happening.

> String.indexOf() is almost certainly a red herring. I 
> suspect the real problem is that you're getting the wrong 
> string. Perhaps you're occasionally seeing docs from a 
> superclass? Or perhaps there's a hash collision?

Those are certainly valid suggestions but I don't think the evidence supports those conclusions.  Hash collision would be equally possible with my change since it's unlikely to affect the hashmap access.

I also don't believe that I'm getting the wrong string because when I saw the failure I saw that the String.indexOf expression returns true when the mText printed did not include @hide or @pending.  I searched because it seemed so implausible.

> You should instrument the code to log the 'needle' and 
> 'haystack' arguments when the expected string is not 
> found. Then run it a hundred times and see what it prints
> when it fails. It won't be the same as the other
> scenarios.

I have done this and the indexOf line returns true that @hide and @pending are present when the mText value is the comment from Intent which does not have a @hide or @pending.

I do not know that 100 builds would be enough since on our builds it does not necessarily show as frequently as this.

My test builds have shown the following statistics:
Builds til problem shows itself: 
7
62
162
231
124
7
36
195
266
74
52

I did get lucky once and get it on the first build I made.

This code has proved to be stable thus far (1000 builds and counting).

From the evidence I have gathered I have come to the surprising conclusion that something does go wrong in indexOf.

Since Google are seeing this issue too, perhaps you could run some tests on this same code to see if I'm right.

I will rerun the tests to see if I can produce a log to confirm that I am not mistaken.

Patch-set: 3