Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Given that the ideal "eo" won't work because the Android locale picker enforces that there must be a country, and given eo_ZZ (ZZ=unknown in ISO 3166) won't work either (because aapt uses ZZ as a sentinel to do pseudolocalization), and given that eo_XX (once suggested for Debian) and eo_001 (suggested at http://www.langtag.net/registries/lsr-region-utf8.txt 001=world) are likely to be less robust across diverse software (XX is user defined and 001 seems unofficial and does not match [A-Z][A-Z]), I agree that eo_EU is the best practical choice here.  So looks good, ship it, and thanks.

Patch-set: 2