Update patch set 2
Patch Set 2:

The -android-linux has deep root in NDK.  I am modifying it but will take a while to fully validate NDK and rebuild the prebuilts.  
As for -Wl,--gc-sections.  I am actually waiting for compiler experts on Cc list to comment.  It's a workaound and I hope final fix can be in by compiler team in binutils-2.21

Patch-set: 2