Update patch set 1
Patch Set 1: Looks good to me, approved

LGTM.  I just wanted to double check with Ying that the length wouldn't be a problem.  Believe it or not, there are a bunch of places in the build where we do hit the commandline limits on Mac OS, which I think are lower than Linux, and have to resort to @files.

Patch-set: 1
Label: Code-Review=+2