Update patch set 4
Patch Set 4: Looks good to me, approved; Verified

This patch won't survive to the next major release (it causes deep merge conflicts), so it's only going to exist in master for a short period. I'm putting it in, and we'll be able to refine it if necessary.

Patch-set: 4
Reviewer: Gerrit User 1000660 <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: SUBM=+1
Label: Verified=+1