Update patch set 1
Patch Set 1:

We have a number of product specifications which all the use same recovery.fstab. Currently we have to either copy the file into the various product directories (and update them all when it changes) or maintain symlinks to the original. Given that the build system allows the user to override the locations of other product definition files (such as the releasetools extensions) it seemed reasonable to do the same here.

Patch-set: 1