Update patch set 1
Patch Set 1:

We are asking for this so that we can organize our product tree in a way that better allows us to support multiple "boards" and "models". Our devices actually have three levels "soc", "board" and "model". By searching a bit deeper, we can put these into a directory hierarchy that actually makes sense from a development and support point of view.

Agreed that 6 may be a bit excessive. Perhaps a smaller number but retaining the concept?

Guilhem... what's the minimum depth that we need for the refactored product tree?

Patch-set: 1