select.mk: make path to ccache externally visible as CCACHE_BIN
Some other Makefiles can make use of this information if they
need to invoke it directly. We change the variable to uppercase
to follow convention of Make variables that are intended to be
accessed outside the scope of the current file.

The immediate use-cases for this are extensions to the build system
to build the Linux kernel in a sub-make, and some of the QEMU
Makefiles which call the compiler in an unusual way.

Change-Id: I597965b89ecb24e7e0208857bb2f7f37192406d0
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Russell Webb <russell.webb@intel.com>