Update patch set 5
Patch Set 5:

I have not been able to push my changes for dalvik and libcore because they are on a specific branch at which I have no permission.
And I have not the time to find how to push those changes.

You can look my changes on my personal repository here :  
https://bitbucket.org/jpilliet/android-292
https://bitbucket.org/jpilliet/libcore-292
https://bitbucket.org/jpilliet/dalvik-292

The problem with the class Enum of libcore is a field accessibility of generic class.
JavaC 6 accepts a code which do not comply the specification.
And JavaC 7 corrects this mistake.
You have just to cast the variable "o" :  
https://bitbucket.org/jpilliet/libcore-292/commits/5074efd3fa79ee6cb0f76f582c4cd1df8b47efff

And You have to modify the file DirectClassFile.java
on dalvik to understand the java 7 classes:  
https://bitbucket.org/jpilliet/dalvik-292/commits/d610d2fe34dbec00f834b0eda5e7dc75d9d16375

Patch-set: 5
Reviewer: Gerrit User 1018065 <1018065@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0