Update patch set 1
Patch Set 1:

So far, it appears that AutoVerifier finds the change to 
be fine and it's asked Ying Wang to review the change.

Perhaps it's a good idea to mention that MIPS/IMGTEC and
Ingenic have been using this patch now for a couple weeks
and we feel satisfied with the change. We are committing
it now so Ingenic and update their device and hardware
code too folks at Google,

-piet

Patch-set: 1