Update patch set 2
Patch Set 2:

There's a reasonably large patchset required to get all host and target modules to build like this globally. I'm not sure if it's something that I should submit or not, so feedback would be welcome.

I felt a need to submit this due to the implications of a bug that writes to a path from root.

Patch-set: 2
Label: Verified=0