Update patch set 1
Patch Set 1:

Thank you, that sounds good. The full patchset to build for target FDO runtime instrumentation globally is reasonably large thus far. However when I clean it up completely, I'll be sure to submit it. This commit should be merged however, due to the implications of the bug it fixes.

In most standard usecases, the developer won't be building with write access to root. However, if they do, or are building inside a chroot, they can quite easily fill up their partition without realizing it after multiple builds for different targets.

Patch-set: 1