Update patch set 1
Patch Set 1:

The Dialer app is newly added in the  4.3 image. The dialer functionality was a part of the Phone app before. Currently, the Dialer app is included in the telephony.mk from 4.3. The telephony.mk file is contained by the full.mk build, but not in the sdk.mk build. Only Dialer app is missed in the telephony.mk, therefore we add it into the sdk.mk directly.

Patch-set: 1
Reviewer: Gerrit User 1006838 <1006838@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0