Update patch set 2
Patch Set 2: Code-Review-2

This was not uploaded to AOSP by the original author, nebkat@teamhacksung.org. It's not clear what the copyright status on this work is, and if we can accept it.

We need the original author to upload this work directly to AOSP him/her self.

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2