Update patch set 1
Patch Set 1:

I am not sure if the difference is significant enough to add a new TARGET_ARCH. Why not make it a TARGET_ARCH_VARIANT of x86? I can see you duplicated most of the x86 config files (the same here) and it will be harder to maintain.

Patch-set: 1