Update patch set 1
Patch Set 1: Code-Review+1

yeah, the documentation seems convincing, plus i forgot that messing with the compiler's include path is how we work around the fact that our "/usr/include" is split into different trees: one for the files we maintain, and the other containing the scrubbed kernel headers.

what we _should_ do is add glibc-like guards to the generated files that check that (for example) linux/socket.h is only #included by <sys/socket.h>. but that would be a different change in a different project.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1