Update patch set 1
Patch Set 1:

I'm not sure this is the right way to make this work. To see an example of how I'm using these new makefiles, see Android.mk in CL:

https://android-review.googlesource.com/#/c/73865/

I don't use all of the variants so I wasn't sure if I should just put what I need or all of the variants.

Patch-set: 1