Update patch set 1
Patch Set 1:

> we shouldn't need libffi for mips64. mips needed it for dalvik, but
 > 64-bit will use art, which doesn't need libffi.

So we should remove libffi after the mips64 port has permanently moved to using only art, and never dalvik.  Currently we have no experience in using art on mips32 or mips64.  Should we rely solely on an untested art when doing initial bringup of mips64?  

I recall there are no plans to extend and tune dalvik for the lp64 environment.  But we should be able to build and run dalvik as a 32-bit process during Android bringup.  For that interim use, retaining libffi a while longer would be helpful.

Patch-set: 1