Update patch set 3
Patch Set 3: Code-Review+1

It is not desirable to truncate it, as long as build fingerprint format is critical for some of the product builds, especially for those that are used for such important milestones as carrier testing, LE, TA, etc. By truncating this we might mask that problem which, in its turn, may cause unneeded delays in a release schedule. Fingerprint format is not as important for developer's builds, but for those we do not fail the build, but provide warning.

Patch-set: 3
Reviewer: Gerrit User 1010605 <1010605@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1