Update patch set 2
Patch Set 2:

I like the idea of this change, because it avoids the need to hardcode revisions in the build system (_u51, _u45 etc.)

If you can find some way to grep for the version numbers in /usr/libexec/java_home - that would be good. I'm not sure it's workable though. On my Mac laptop (running Mavericks) it points to ..../JavaVM.Framework/Versions/Current/ which is itself a symlink to another location.

I'll be happy to accept this change if you can get it to work !

Patch-set: 2