Update patch set 1
Patch Set 1: Code-Review-1

Not sure we want this. its sort of a feature that we run with the debug builds for all the tests. if anything, I'd try to find a way to not build the non-debug versions. We have some Android.mk options for this, we could set to only build host debug. We'd have to verify that those options work, though I believe Intel submitted a patch to fix one case recently.

Patch-set: 1