atree: Handle DT_UNKNOWN in readdir with stat properly
Operating systems that support d_type in readdir may still not
be able to provide any information in d_type because of the
underlaying filesystem. Let this case also fall back to stat.
Also, the stat call in files.cpp was using the wrong path and
never worked. Fixed this and check the return code of stat.

Change-Id: I7d4d66a415780a4dffd9313a88b28570f81367e0