Update patch set 1
Patch Set 1:

Look at external/gtest/src/Android.mk, it actually depends on the NDK's stlport_static, so it can be used in unbundled build. It doesn't restrict the undefined symbols to be resolved from libstdc++.a or libstdc++.so. In our case it happens the broken module requires LOCAL_FORCE_STATIC_EXECUTABLE.

Patch-set: 1