Update patch set 1
Patch Set 1:

No, these are two different things.  We need to know whether or not a target has support for 64-bit apps (meaning it has a 64-bit zygote installed).  We also need to know for the cases where it supports both 32-bit and 64-bit apps whether which one to target if both are possible.

Patch-set: 1