executable_prefer_symlink: add common makefile for creating symlink to prefer binary version
For a 64bit target, we can have support for both 32bit and 64bit version binary(like app_process, dalvikvm),
and there are some commands like am/pm will call these commands to implement it's function.
In some targets it will prefer to use the 32bit version, and in other targets it will prefer to use the 64bit version.

So this change is to add the common work for creating the symlink to the prefer version binary,
and avoid the change on the caller commands side like am/pm

In this change, it uses TARGET_PREFER_32_BIT_APPS to decide linking to 32bit version or 64bit version.

Change-Id: I96f041c0e0ededd953feead30ec63ad63e945a6f
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>