Update patch set 2
Patch Set 2:

> I'm not sure what dalvikvm is used for from the command line.

dalvikvm is used by CTS vm-tests. but we are going to need to be able to run both 32-bit and 64-bit. that said, I think we could use it there as many other internal and external test tools that run dalvikvm will probably expect it to run the "primary" runtime.

Patch-set: 2