New target products for base Android system images to be run under QEMU.
This patch adds 5 new target products that can be used to build very
minimal system images that should be runnable under upstream QEMU
once it has been patched to support a few Android-specific virtual
devices and boards.

Note that this is different from running under the Android emulator.
While the images currently run under it, this might break in the future.

The main point here is the intent to send patches to upstream QEMU
to support emulation of Android system images. These patches will
likely only be accepted only if there is a way for the upstream
team to check that the changes they perform to their source code
doesn't break Android.

In other words, the corresponding system image binaries are intended
to be freely distributable to ease upstream QEMU testing. Note that
they should, at a minimum, support ADB and the ability to install
and use executables and libraries under /data/local/tmp/ or
similar directories.

To build and test the system images, do something like:

   . build/envsetup.sh
   lunch qemu_arm-userdebug
   make -j8
   emulator -shell

This should give your a root shell.

Change-Id: Ie86c2a3ffdffdc8208fd660cb8bb5de3a31e01ab