Update patch set 1
Patch Set 1: Verified-1

Note: I haven't check a Clang-based build with this patch, what is the recommend way to test this case?

Also, I believe that rebuilding the emulator from sources with the platform build (i.e. by defining BUILD_EMULATOR=true in your environment before 'make') will generate a broken binary (i.e. one that doesn't boot a system image, reasons unclear).

I'll try to prepare a patch to external/qemu that enforces the use of the 4.6 toolchain in the mean time. Please do not submit this until this is submitted. That's why I'm putting a -1 for now here.

Patch-set: 1
Label: Verified=-1