Update patch set 1
Patch Set 1:

This is bound to raise some eyebrows...

This may not be the right solution, but I'm uploading it to start the discussion on what a better solution might look like if this one isn't it.

Was this something we wanted to have applied across the board? git blame shows that these lines hadn't been modified since the initial import, so it's hard to say why they were originally added.

I'm looking into some benchmarks and checking file sizes right now to see if this is something we should definitely avoid.

Patch-set: 1