Update patch set 1
Patch Set 1:

Any size differences are minuscule (not significant enough to show a change when using du -sh on libart.so, libc.so, or system.img).

Running the framework benchmarks show negligible differences.

More thorough testing might show that sizes are slightly larger and performance is slightly worse with this, but it looks insignificant at this point.

Patch-set: 1