Update patch set 1
Patch Set 1:

Why is there never a simple answer to some questions ...

Not quite future proofing, the default already is to use the User space logger unless specified otherwise (in an associated product board.mk), and I neglected the emulator as a 'product' as I phased in the user space logger (logd). I do not use the emulator, so my WAG is that they (you know, them are they) have been missing the logger for some time now(?) and never reported it to me.

The future-proofing angle is that I expect this and the user space logger to be demonstrably running and tested on the product (I have individually tested N4, N5, 2xN7, N10 and development products) *before* I turn off the kernel-based logger. I did this phased in about two months after.

This change remains 'untested' and no 'build' testing (yet) as I will also have to adjust the product-associated init.rc to start logd to really enable it, it does no harm to have logd present on or off (unless it fails to build properly for the emulated environment) as I gear up to eventually test and/or solve the user space logger in the emulation environment.

Sight unseen (meaning I view this as a low-risk adjustment that may not require direct unit testing/verification), this is the first groundwork shot of possibly several logd related configuration adjustments.

Patch-set: 1