Update patch set 1
Patch Set 1: Verified+1

Tons of projects are failing to build with this. The AOSP build server shows constant breakages in nearly every target after this was merged. I reverted this locally and it fixed the problem.

Ying: I'm not really sure why exactly this is happening. Looking at frameworks/compile/libbcc, the issue is that a bunch of compile flags are being passed in the link step, causing clang to throw warnings, which get turned in to errors by -Werror.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1