Add support building SDK add-on package without a system image.
Currently to build the SDK add-on includes building a whole system image
for emulator. But in many scenarios, the add-on feature is provided by
specific HW and does not apply to the emulator. So to create a simple
add-on package is enough without a system image. The usage is to specify
"PRODUCT_SDK_ADDON_LIBRARY_ONLY := true" into the feature_addon.mk.

Change-Id: I22d3c899223b98035add94188d65dd85c4049f55
Signed-off-by: Jackie Wu <jackie.wu@intel.com>
Signed-off-by: Wenda Song <wenda.song@intel.com>