Add domains for goldfish services.
goldfish-setup, goldfish-logcat, and qemu-props are goldfish-specific
oneshot services that lacked domain definitions and thus were left in init's
domain.

For goldfish-logcat, rather than introducing a separate exec type for
the logcat program just for this purpose, we instead rely on a seclabel
entry in the init.goldfish.rc file to assign the domain for the service.

Change-Id: If52a0a891a9ee9e7d27cd705b7b2ae8990cf7924
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>