Add domains for goldfish services.
goldfish-setup, goldfish-logcat, and qemu-props are goldfish-specific
oneshot services that lacked domain definitions and thus were left in init's
domain.

For goldfish-logcat, rather than introducing a separate exec type for
the logcat program just for this purpose, we instead rely on a seclabel
entry in the init.goldfish.rc file to assign the domain for the service.

Change-Id: Idce1fb5ed9680af84788ae69a5ace684c6663974
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>