Add option for generating coverage info.
Set `LOCAL_NATIVE_COVERAGE := true` to generate coverage information for
a given component. The environment variable `NATIVE_COVERAGE=true` can
be set to enable coverage for all (target, non-clang) native code in the
tree. Any module explicitly setting `LOCAL_NATIVE_COVERAGE := false`
will override this behavior. Note that a clean build is required when
flipping the value of this flag.

This is currently not supported for clang (b/17574078, b/17583330).

If static library A is included in a binary B (dynamic or static
executable, or shared library), and A is built with coverage
information, B is required to link with libgcov.a. Since the make does
not offer a good way to track this dependency, unconditionally link
libgcov.a. This ensures that all of the libgcov dependencies will always
be resolved, and causes no change in the resulting binary if coverage is
not used.

To get coverage information from the device:

1. sudo apt-get install lcov
2. Build with coverage information.
3. Push the new binaries to the device.
4. Run tests with the additional environment variables:
    * GCOV_PREFIX=/data/local/tmp/gcov
    * GCOV_PREFIX_STRIP=`echo $(ANDROID_BUILD_TOP) | grep -o / | wc -l`
5. Run `acov`.

acov will pull all coverage information from the device, push it to the
right directories, run lcov, and display the coverage report (currently
by opening it in your browser).

Bug: 10134489
Change-Id: Id5a19f2c215e4be80e6eae27ecc19b582f2f6813