Update patch set 1
Patch Set 1:

I am not if adding -fPIC to ASFLAGS is sufficient.  On GCC (as part of a google/android local change) the -fPIC flag is set by default for GCC and this causes it to send -KPIC to the assembler.  But it also causes GCC itself to generate different code than if -fPIC was not on and it may change the link line as well.

If using the -fPIC flag on Clang does more then just pass -KPIC to the assembler you want those other actions to happen as well.

Patch-set: 1
Reviewer: Gerrit User 1050199 <1050199@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0