Update patch set 1
Patch Set 1:

So, one place we'll run in to problems with this is when using a static library built with clang (and thus with ASAN) gets used in a binary built with GCC. This is probably a pretty rare circumstance, and we can just set LOCAL_ADDRESS_SANITIZER := false for those static libs (or maybe move those projects to clang?). I'll know more once I make it through a whole build.

Patch-set: 1