Update patch set 1
Patch Set 1:

This should not be on by default in eng builds today. Let's experiment with this a bit for now, and see how well it works in practice for our current clang projects. There is also the remaining issue that ASan currently crashes in AOSP, so we should get that fixed.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0