Update patch set 6
Patch Set 6:

Verified build in AOSP.

The prebuilts have been checked in for about a day with no reports of breakage. Checking this change in now. It should effectively be a no-op since the two libraries are still the same.

Once this is in, and no breakages have been reported, I'll rebase the CL to update the protobuf library to 2.6 and send the announcement out to nano-protos (bcc:android-eng) with migration info.

Patch-set: 6