Make modules depend on their makefile.
This should obviate much of the need for cleanspecs, and also make it
unnecessary to continue adding LOCAL_ADDITIONAL_DEPENDENCIES for this
sort of thing all over the tree.

This will also make each module depend on the build system makefiles
it passes through on the way to each location.

Change-Id: I97aa8fd280ae868a5f6364f8b7bf3c2fe235d6ce