Update patch set 1
Patch Set 1:

I would love a way to make each target depend on the command line that builds it, rather than having to depend on the makefile that defines it (since that means each target also depends on the build flags of the other modules in the file...). Ninja does this, but I don't think there is a way to work it in to our system without significant effort.

Patch-set: 1