Update patch set 2
Patch Set 2:

Okay, this one is a little less aggressive (and actually behaves properly). Turns out I can't rely on the original makefile being the second to last element in MAKEFILE_LIST, but apparently there's a LOCAL_MODULE_MAKEFILE that does what I need it to.

SInce LOCAL_MODULE_MAKEFILE is updated on a call to my-dir, this won't work properly if someone includes a makefile that calls my-dir, then defines more modules in the outer makefile. There are other issues with this pattern, so if there are uses of this we should just fix them.

Patch-set: 2