Update patch set 1
Patch Set 1:

As I said in the comment, it can be other arch variants. The name may also change in the future. To check if it's tapas build or not, you are already dependent on the build environment, against your purpose of "make lunch target irrelevant". So if you want to support gdb_client for binaries built by tapas, it won't be worse to get  $DEVICE from the build system.

Patch-set: 1