Update patch set 3
Patch Set 3:

Andrew, if I'm reading that bug correctly, the problem was actually a broken prebuilt that needed to be updated. If that's the case, this should be safe to submit (though any broken prebuilts will still need to be fixed, and there isn't a good way to flush those out aside from submitting and looking for smoke).

Does that make sense, or is there something else I'm missing here?

Patch-set: 3