Generate odex in data partition
With L, odex files are binaries which are bigger than odex on K.
These ones increase heavily the size of the system partition
when using WITH_DEX_PREOPT := true.

We add a new way to generate odex files in order to store them
in the data partition and reduce the system size partition as
those one will be wiped during an update of the system application.
Ideally, all updateable applications should be in the /data partition

We add the DEX_PREOPT_IN_DATA variable to enable the copy
of odex files in the proper directory in data.

The list of applications to push in /data is selected by
setting DEX_PREOPT_IN_DATA_LIST.
The ones not figuring in this list will stay in /system.
If this list is empty the behavior stay the same as it was
with WITH_DEX_PREOPT := true only.

As we will need at some point to generate again these files,
we keep the dex files in the apk but as these files are smaller
than odex we still gain on that side.

This new option add the need to flash userdata partition and
increase boot time during a factory reset which can be mitigated
depending on the applications kept in the /system partition.

Change-Id: I9e5742193a24945549b092a69cc8a136e16a5d8c
Signed-off-by: Julien Delayen <julien.delayen@intel.com>
Signed-off-by: Guilhem Imberton <guilhem.imberton@intel.com>