Generate odex in data partition
With L, odex files are binaries which are bigger than odex on K.
These ones increase heavily the size of the system partition
when using WITH_DEX_PREOPT := true.

We add a new way to generate odex files in order to store them
in the data partition and reduce the system size partition as
those one will be wiped during an update of the system application.
Ideally, all updateable applications should be in the /data partition

We add the DEX_PREOPT_IN_DATA variable to enable the copy
of odex files in the proper directory in data.

The list of applications to push in /data is selected by
setting DEX_PREOPT_IN_DATA_LIST.
The ones not figuring in this list will stay in /system.
If this list is empty the behavior stay the same as it was
with WITH_DEX_PREOPT := true only.

As we will need at some point to generate againthese files,
we keep the dex files in the apk but as these files are smaller
than odex we still gain on that side.

We create a file .preload in each instruction set dalvik cache
folder to indicate that this one has been preloaded. This
information is then used by art to avoid pruning the dalvik-cache
during the first boot.

This new option add the need to flash userdata partition and
increase boot time during a factory reset which can be mitigated
depending on the applications kept in the /system partition.

Change-Id: I915a90d80fd5fbd2f4d6d45e4a585127c80e2e50
Depends-On: I9e5742193a24945549b092a69cc8a136e16a5d8c
Depends-On: I6c49d7028ee118ee21bf786d7b0e1c955b722d29
Depends-On: Ia7a736c454f05daf3fb3822b8a2f6604fe531de0
Depends-On: Ie01be98fe39107716b43bfba384d43b6ca6b81d0
Depends-On: I3ccb040cc937117745fa798c20be323f2ced943f
Signed-off-by: Julien Delayen <julien.delayen@intel.com>
Signed-off-by: Guilhem Imberton <guilhem.imberton@intel.com>