Update patch set 2
Patch Set 2: Code-Review+1

Ah, I see. I'm surprised, because we actually had a bug opened recently complaining that Darwin was still using libstdc++ 4.2 instead of libc++.

Chromium should be the only thing broken by this, and they have a different approach in the works that should make this unnecessary, but I don't see a problem with adding the work around for the time being. (We'll want to keep the mac_version.mk change in the long run, but won't need the combo changes.)

Patch-set: 2
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1