Update patch set 1
Patch Set 1:

> I've experimented with this before -- the only bigger problem was
 > that some older parts of the code typedef-ed char16_t, causing the
 > ABI to mismatch ("real" char16_t references generated by C++11
 > code, references to the typedef generated by C++98 code).

I've dealt with this as well. I'm in the process of uploading a whole bunch of CLs to make this patch safe to submit.

 > I wonder if it wouldn't make more sense to go straight to
 > -std=c++1y (AKA c++14) though... It's supported by both current gcc
 > and current clang, and it's mostly a set of fixes to c++11 anyway.
 > 
 > Binary literals (at last!) and relaxed constexpr restrictions could
 > be quite useful.

One step at a time :) I don't think the version of GCC we're using has full support yet, and I'd like to wait until then.

Patch-set: 1