Fix dependencies for NDK builds.
NDK builds still need the libraries in LOCAL_SYSTEM_SHARED_LIBRARIES
to be built. This problem shows up when running mma in external/skia
from a clean build because skia doesn't depend on libstdc++, but it
does depend on libexpat, which does and is built with the NDK. When
libexpat is built, it uses the libstdc++ in prebuilts/ndk, but never
builds libstdc++, so the linker can't find it when linking skia.

Change-Id: I28e8212c8c3814be66b526735873507eebbb6edb