Update patch set 1
Patch Set 1:

Okay, after I fixed my Xcode issue on the mac, I found a different bug unrelated to my change. This change works just fine on the mac for target and host static libraries.

The bug is that if transform-host-o-to-static-lib is called with a library that doesn't include any object files (because it's a shell library using whole static libraries to create it), you get an error from ar because it has nothing to do. This works for the host libraries on my linux system but not on my mac, so I'm not sure this is worth pursuing.

So as far as I can tell, everything is working on both mac and linux.

Patch-set: 1