Introduce LOCAL_ZIPALIGN option (false)
By default the build system zip aligns PRESIGNED modules,
however, some of these PRESIGNED apks can contain a tampering
mechanism that involves a signature of the apk itself.
Therefore, to re-zipalign again the apk will make fail the apk
due to its signature has been modified (an example of this
is the tampering mechanism used by DexGuard).

This property allows to skip zipaligning in a PRESIGNED apk
if desired.

Change-Id: I06cbfa7b4f2b8d382c2dd93d0c8165374bad8ee9